refactor: share deterministic analysis with browser callers
